next-word-prediction/requirements.txt
azure-functions
numpy
transformers
http://download.pytorch.org/whl/cpu/torch-1.13.0%2Bcpu-cp39-cp39-linux_x86_64.whl

